Break Model.cc into Model + Modules
